% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc.R
\name{fcc_broadband_by_state}
\alias{fcc_broadband_by_state}
\title{Count broadband providers by state}
\usage{
fcc_broadband_by_state(vintage = "dec2020")
}
\arguments{
\item{vintage}{Filing vintage (default "dec2020")}
}
\value{
tibble: stateabbr, provider_count
}
\description{
Count broadband providers by state
}
