% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc.R
\name{fcc_csric}
\alias{fcc_csric}
\title{CSRIC communications security best practices}
\usage{
fcc_csric(keyword = NULL, network_type = NULL, limit = 500)
}
\arguments{
\item{keyword}{Search keyword in description}

\item{network_type}{Filter: "Cable", "Wireless", "Wireline", "Satellite",
"Internet/Data"}

\item{limit}{Max rows (default 500)}
}
\value{
tibble
}
\description{
Searchable database of best practices from the Communications Security,
Reliability and Interoperability Council (CSRIC).
}
