% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc.R
\name{fcc_pra}
\alias{fcc_pra}
\title{FCC information collections under the Paperwork Reduction Act}
\usage{
fcc_pra(q = NULL, limit = 1000)
}
\arguments{
\item{q}{Full-text search term}

\item{limit}{Max rows (default 1000)}
}
\value{
tibble
}
\description{
Active and pending information collections for which FCC has OMB approval.
}
