% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcc.R
\name{fcc_psap}
\alias{fcc_psap}
\title{911 Public Safety Answering Point (PSAP) registry}
\usage{
fcc_psap(state = NULL, q = NULL, limit = 1000)
}
\arguments{
\item{state}{Two-letter state abbreviation}

\item{q}{Full-text search term}

\item{limit}{Max rows (default 1000)}
}
\value{
tibble
}
\description{
Master list of PSAPs (911 call centers) in the US.
}
